Add lib/autoload.php to allow manual (i.e. without composer) installation.
The custom autoloader created by this commit translates class
names into filesystem locations by a simple rule: it replaces
backslashes with frontslashes, and adds ".php" on the end. Two
locations are searched for this class implementation file: first we
check under lib/, then under test/.

The existing test/bootstrap.php checked for vendor/autoload.php and
would show an error to the user if it did not exist. Now that we have
our own autoloader, we can just use that instead. So, test/bootstrap.php
was removed, and phpunit.xml.dist was pointed at the new lib/autoload.php
for its bootstrap code.

Thanks to Artur Weigandt for the implementation of the autoloader.